move ts bindings into their own folder
