+++ crypto/bindings.ts
@@ -4,11 +4,11 @@
  * It gets imported as the first thing in ../../snarky.js so that the global variable is ready by the time JSOO code gets executed.
  */
 import { prefixHashes, prefixHashesLegacy } from '../crypto/constants.js';
-import { Bigint256Bindings } from './bindings-bigint256.js';
-import { PallasBindings, VestaBindings } from './bindings-curve.js';
-import { FpBindings, FqBindings } from './bindings-field.js';
-import { FpVectorBindings, FqVectorBindings } from './bindings-vector.js';
-import { createRustConversion } from './bindings-conversion.js';
+import { Bigint256Bindings } from './bindings/bigint256.js';
+import { PallasBindings, VestaBindings } from './bindings/curve.js';
+import { FpBindings, FqBindings } from './bindings/field.js';
+import { FpVectorBindings, FqVectorBindings } from './bindings/vector.js';
+import { createRustConversion } from './bindings/conversion.js';
 
 const tsBindings = {
   prefixHashes,
