+++ crypto/bindings/util.ts
